Exercises 4.18-4.22

+++ controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/Blog');
-const User = require('../models/User');
 
 blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({})
@@ -12,47 +11,64 @@ blogsRouter.get('/', async (req, res) => {
 });
 
 blogsRouter.post('/', async (req, res) => {
-  if (!req.body.title || !req.body.url) {
-    return res.status(400).json('error: title or url missing');
-  }
-  if (!req.body.likes) {
-    req.body.likes = 0;
-  }
+  if(!req.user)
+    return res.status(401)
+      .json({ error: 'token missing or invalid' });
 
-  const user = await User.findById(req.body.userId);
+  if (!req.body.title || !req.body.url) 
+    return res.status(400)
+      .json('error: title or url missing');
+
+  if (!req.body.likes) req.body.likes = 0;
+
   const blog = new Blog({
     ...req.body,
-    user: user.id
+    user: req.user.id
   });
+
   const returnedBlog = await blog.save();
-  user.blogs = user.blogs.concat(returnedBlog._id);
-  await user.save();
+  req.user.blogs = req.user.blogs.concat(returnedBlog._id);
+  await req.user.save();
 
   res.status(201).json(returnedBlog);
 });
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const blog = await Blog.findById(req.params.id);
+  if (!req.user || blog.user.toString() !== req.user.id.toString())
+    return res.status(401)
+      .json({ error: 'token missing or invalid' });
+
   const updatedBlog = await Blog.findByIdAndRemove(req.params.id);
   res.status(204).json(updatedBlog);
 });
 
 blogsRouter.put('/:id', async (req, res) => {
-  if (!req.body.title || !req.body.url || !req.body.likes) {
-    return res.status(400).json('error: title or url missing');
-  }
+  const blog = await Blog.findById(req.params.id);
+  if (!req.user || blog.user.toString() !== req.user.id.toString())
+    return res.status(401)
+      .json({ error: 'token missing or invalid' });
+
   const body = req.body;
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
+
+  if (!body)
+    return res.status(400)
+      .json('error: info missing');
+
+  const updatedBlog = {
+    title: body.title || blog.title,
+    author: body.author || blog.author,
+    url: body.url || blog.url,
+    likes: body.likes || blog.likes,
   };
-  const updatedBlog = await Blog
-    .findByIdAndUpdate(
-      req.params.id, 
-      blog, 
-      { new: true, runValidators: true });
-  res.status(200).json(updatedBlog);
+
+  const returnedBlog = await Blog.findByIdAndUpdate(
+    req.params.id, 
+    updatedBlog, 
+    { new: true, runValidators: true }
+  );
+
+  res.status(200).json(returnedBlog);
 });
 
 module.exports = blogsRouter;
